keeps cookies, changes window on signup/login, moved login to seperate component
