+++ src/components/app/App.jsx
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
 import Home from '../../Pages/Home';
+import Login from '../../Pages/Login';
 import Header from '../../Pages/Header';
 import AboutUs from '../../Pages/AboutUs';
 import Submit from '../../Pages/Submit';
@@ -28,6 +29,12 @@ class App extends Component {
         <Switch>
 
           <Route path="/" exact={true}
+            render={routerProps => (
+              <Login {...routerProps} />
+            )}
+          />
+
+          <Route path="/home/" exact={true}
             render={routerProps => (
               <Home {...routerProps} />
             )}
